chore: implement Ansgar's solution for licenses through Projen
This brings the license generation into the Projen template to
ensure that it does not get overwritten in the future, copying
Ansgar's implementation from the AWS Adapter. Note that this
project is currently using an Apache-2.0 license where HashiCorp's
standard has been MPL-2.0 and I don't know if using Apache here is
intentional or an accidental oversight. But, we can always fix
that in a follow-up PR.