chore: add issue templates (#50)
As we're starting to see more activity in this repo, especially
community contributors coming in, I figured these are good to have.

FWIW I looked into whether there is a Projen-native way to generate
issue templates and there's not (there only is for the pull request
template), and translating our existing YAML templates into JavaScript
so that Projen can compile it back into YAML seems silly. So, just
adding in the YAML files and having them not be managed by Projen at all
seemed like the best solution. IMO this also makes them easier to
maintain in the future (for example, if we want to add a new field to an
issue template in all our repos). But if anyone disagrees with this
approach, let me know.