chore: automate updating the README (#55)
This uses the cool `action-docs` package to automagically generate the
inputs in the readme based on the `action.yml` file (which in turn is
managed by Projen). This alleviates us from needing to manually update
the readme whenever we add inputs, change the default value, etc.

This goes part of the way to addressing #53 -- if we figure out a way to
automagically update the CDKTF (and Terraform) version numbers in
`.projenrc.ts` when new versions are released, then we're very close to
having a seamless automated process that we can easily step away from as
needed.