chore!: change default Terraform version to 1.7.0 (#138)
This PR increases the default version of Terraform used from `1.6.6` to
version `1.7.0`.
This is considered a breaking change because anyone who does not
manually specify a `terraformVersion` in their action configuration will
automatically start using the new version.
 
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update the _Terraform Version_ to `1.7.0` in the TFC web UI for
the
[cdk-action-testing](https://app.terraform.io/app/cdktf/workspaces/cdk-action-testing/settings/general)
workspace
 
Please complete the above steps and then mark this PR as ready for
review to rerun the checks. Thanks!

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen Kwon <855115+xiehan@users.noreply.github.com>