chore!: Require `setup-terraform` to be run before action (#255)
<!--

Unless this is a very simple 1-line-of-code change, please create a new
issue describing the change you're proposing first, then link to it from
this PR.

Read more about our process in our contributing guide:
https://github.com/hashicorp/terraform-cdk-action/blob/main/.github/CONTRIBUTING.md

-->

### Related issue

Fixes #243 

### Description

This PR makes it so that this action doesn't install Terraform
automatically. The expectation is that terraform with the specific
version is already installed on the runner. This can be done with the
[setup-terraform
action](https://github.com/marketplace/actions/hashicorp-setup-terraform).
The examples are updated to explain the usage.

### Checklist

- [ ] I have updated the PR title to match [CDKTF's style
guide](https://github.com/hashicorp/terraform-cdk-action/blob/main/.github/CONTRIBUTING.md#pull-requests-1)
- [ ] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works if applicable
- [ ] New and existing unit tests pass locally with my changes

<!-- If this is still a work in progress, feel free to open a draft PR
until you're able to check off all the items on the list above -->

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Ansgar Mertens <ansgar@hashicorp.com>